+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Created `tar_stars()` and `tar_stars_proxy()` that create `stars` and `stars_proxy` objects, respectively.
 * Created `tar_terra_tiles()`, a "target factory" for splitting a raster into multiple tiles with dynamic branching.
+* Removes the `iteration` argument from all `tar_*()` functions.  `iteration` now hard-coded as `"list"` since it is the only option that works (for now at least).
 * Created two helper functions for use in `tar_terra_tiles()`: `tile_grid()` and `tile_blocksize()`.
 * Created utility function `set_window()` mostly for internal use within `tar_terra_tiles()`.
 * Added the `description` argument to all `tar_*()` functions which is passed to `tar_target()`.